+++ CHANGELOG.md
@@ -1,9 +1,21 @@
 # Change Log
 
+## 2023-02-02
+- Vue2 version 1.0.7 published
+- Vue3 version 1.0.6 published
+- React version 1.0.4 published
+- update icons
+
+## 2023-01-31
+- Vue2 version 1.0.6 published
+- Vue3 version 1.0.4 published
+- React version 1.0.3 published
+- update icons
+
 ## 2022-10-09
 - Vue2 version 1.0.5 published
 - Vue3 version 1.0.3 published
-- React vertion 1.0.2 published
+- React version 1.0.2 published
 - update icons
 
 ## 2022-10-08